+++ source/ignore_files_based_on_metadata/plugin.py
@@ -81,14 +81,36 @@ def file_has_disallowed_metadata(path, disallowed_metadata, metadata_value):
 
     # Check if stream or format components contain disallowed metadata
     streams = [probe_streams[i] for i in range(0, len(probe_streams)) if "codec_type" in probe_streams[i] and probe_streams[i]["codec_type"] == "video"]
-    processed_streams_already = [streams[i] for i in range(0, len(streams)) if disallowed_metadata in streams[i] and metadata_value in streams[i][disallowed_metadata]]
+    file_has_disallowed_metadata = [streams[i] for i in range(0, len(streams)) if disallowed_metadata in streams[i] and metadata_value in streams[i][disallowed_metadata]]
     probe_format_d = {k:v for  (k, v) in probe_format.items() if type(v) is dict}
     probe_format_kv = {k:v for  (k, v) in probe_format.items() if type(v) is not dict}
     for v in probe_format_d.values():
         probe_format_kv.update(v)
-    processed_streams_already2 = [(k, v) for (k, v) in probe_format_kv.items() if (disallowed_metadata in k.lower() and metadata_value in v)]
+    file_has_disallowed_metadata_fmt = [(k, v) for (k, v) in probe_format_kv.items() if (disallowed_metadata in k.lower() and metadata_value in v)]
 
-    if processed_streams_already or processed_streams_already2:
+    # Check if video, audio, or attachement stream tags contain disallowed metadata
+    attachment_streams = [probe_streams[i] for i in range(0, len(probe_streams)) if "codec_type" in probe_streams[i] and probe_streams[i]["codec_type"] == "attachment"]
+    try:
+        probe_as_tags_kv = {k:v for i in range(0, len(attachment_streams)) for (k, v) in attachment_streams[i]["tags"].items() if type(v) is not dict}
+        file_has_disallowed_metadata_ast = [(k, v) for (k, v) in probe_as_tags_kv.items() if (disallowed_metadata in k.lower() and metadata_value in v)]
+    except KeyError:
+        file_has_disallowed_metadata_ast = ""
+
+    video_streams = [probe_streams[i] for i in range(0, len(probe_streams)) if "codec_type" in probe_streams[i] and probe_streams[i]["codec_type"] == "video"]
+    try:
+        probe_vs_tags_kv = {k:v for i in range(0, len(video_streams)) for (k, v) in video_streams[i]["tags"].items() if type(v) is not dict}
+        file_has_disallowed_metadata_vst = [(k, v) for (k, v) in probe_vs_tags_kv.items() if (disallowed_metadata in k.lower() and metadata_value in v)]
+    except KeyError:
+        file_has_disallowed_metadata_vst = ""
+
+    audio_streams = [probe_streams[i] for i in range(0, len(probe_streams)) if "codec_type" in probe_streams[i] and probe_streams[i]["codec_type"] == "audio"]
+    try:
+        probe_aus_tags_kv = {k:v for i in range(0, len(audio_streams)) for (k, v) in audio_streams[i]["tags"].items() if type(v) is not dict}
+        file_has_disallowed_metadata_aust = [(k, v) for (k, v) in probe_aus_tags_kv.items() if (disallowed_metadata in k.lower() and metadata_value in v)]
+    except KeyError:
+        file_has_disallowed_metadata_aust = ""
+
+    if file_has_disallowed_metadata or file_has_disallowed_metadata_fmt or file_has_disallowed_metadata_ast or file_has_disallowed_metadata_vst or file_has_disallowed_metadata_aust:
         logger.debug("File '{}' contains disallowed metadata '{}': '{}'.".format(path, disallowed_metadata, metadata_value))
         return True