+++ source/ignore_files_based_on_metadata/changelog.md
@@ -1,3 +1,7 @@
+
+**<span style="color:#56adda">0.0.3</span>**
+- add test to check attachment streams tags
+
 **<span style="color:#56adda">0.0.2</span>**
 - fix __init__.py filename